Moved getRadiusFromPixel to Emitter class.
